+++ x/valset-pref/export_test.go
@@ -2,9 +2,11 @@ package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	"github.com/osmosis-labs/osmosis/osmomath"
 	lockuptypes "github.com/osmosis-labs/osmosis/v20/x/lockup/types"
+	"github.com/osmosis-labs/osmosis/v20/x/valset-pref/types"
 )
 
 type (
@@ -13,4 +15,13 @@ type (
 
 func (k Keeper) ValidateLockForForceUnlock(ctx sdk.Context, lockID uint64, delegatorAddr string) (*lockuptypes.PeriodLock, osmomath.Int, error) {
 	return k.validateLockForForceUnlock(ctx, lockID, delegatorAddr)
+}
+
+func (k Keeper) GetValsetRatios(ctx sdk.Context, delegator sdk.AccAddress,
+	prefs []types.ValidatorPreference, undelegateAmt osmomath.Int) ([]ValRatio, map[string]stakingtypes.Validator, osmomath.Dec, error) {
+	return k.getValsetRatios(ctx, delegator, prefs, undelegateAmt)
+}
+
+func FormatToValPrefArr(delegations []stakingtypes.Delegation) []types.ValidatorPreference {
+	return formatToValPrefArr(delegations)
 }